Semana12 - Texturas e Iluminación/requirements.txt
numpy==1.21.6
pygame==2.4.0
PyOpenGL==3.1.7
